Optimization & Update README
- 修复了 BuildSF 接口bug.
- 更新 README

+++ src/ctp_frame.c
@@ -12,7 +12,7 @@
 
 
 uint8_t 
-Cus_Cantp_BuildSF( uint8_t *frame, Cus_CANTP_FrameSize_t frameSize, uint8_t pciOffset, const uint8_t *data, uint8_t len )
+Cus_Cantp_BuildSF( uint8_t *frame, Cus_CANTP_FrameSize_t frameSize, uint8_t pciOffset, const uint8_t *data, uint32_t len )
 {
 	if ( !frame || !data )
 		return 0;
